+++ mayan/apps/acls/classes.py
@@ -27,6 +27,22 @@ class ModelPermission:
         cls._model_permissions.pop(model, None)
 
     @classmethod
+    def get_choices_for_class(cls, klass):
+        result = []
+
+        for namespace, permissions in itertools.groupby(cls.get_for_class(klass=klass), lambda entry: entry.namespace):
+            permissions = [
+                (permission.pk, permission) for permission in permissions
+            ]
+            result.append(
+                (namespace, permissions)
+            )
+
+        # Sort by namespace label.
+        result.sort(key=lambda entry: entry[0].label)
+        return result
+
+    @classmethod
     def get_classes(cls, as_content_type=False):
         ContentType = apps.get_model(
             app_label='contenttypes', model_name='ContentType'
@@ -49,28 +65,18 @@ class ModelPermission:
         return cls._field_query_functions[model]
 
     @classmethod
-    def get_for_class(cls, klass, as_choices=False):
-        if as_choices:
-            results = []
+    def get_for_class(cls, klass):
+        # Return the permissions for the klass and the models that
+        # inherit from it.
+        result = set()
+        result.update(cls._model_permissions.get(klass, ()))
+        for model in cls._inheritances_reverse.get(klass, ()):
+            result.update(cls._model_permissions.get(model, ()))
 
-            for namespace, permissions in itertools.groupby(cls.get_for_class(klass=klass, as_choices=False), lambda entry: entry.namespace):
-                permission_options = [
-                    (force_text(s=permission.pk), permission) for permission in permissions
-                ]
-                results.append(
-                    (namespace, permission_options)
-                )
+        result = list(result)
+        result.sort(key=lambda permission: permission.namespace.name)
 
-            return results
-        else:
-            # Return the permissions for the klass and the models that
-            # inherit from it.
-            result = []
-            result.extend(cls._model_permissions.get(klass, ()))
-            for model in cls._inheritances_reverse.get(klass, ()):
-                result.extend(cls._model_permissions.get(model, ()))
-
-            return result
+        return result
 
     @classmethod
     def get_for_instance(cls, instance):