+++ HISTORY.rst
@@ -330,6 +330,8 @@
 - Remove trailing new lines from the MIME type and encoding returned by the
   ``MIMETypeBackendFileCommand``.
 - Make ``MIMETypeBackendFileCommand`` the default MIME backend.
+- Fix sorting and grouping of permissions in the workflow action to grant
+  or revoke document access.
 
 4.2.5 (2022-05-21)
 ==================